refactor(character): rename brief generation module
