+++ application/characters/generate_briefs.py
